+++ tests/testthat/test_careless_inattentive.R
@@ -1,19 +1,20 @@
 ################################################################################
 # setup
 set.seed(26112020)
+# generating items' parameters - "simultaenous" mode
 nItems <- 20
 sMSml <- make_scoring_matrix_aem(1:5, "simultaneous")
-slopesSml <- matrix(c(rep(10, nItems),
-                      rlnorm(nItems, 0, 0.3),
-                      rep(1, nItems),
-                      rep(1, nItems),
-                      rep(1, nItems)),
-                    ncol = 5, dimnames = list(NULL, c("ci", colnames(sMSml))))
-interceptsSml <- runif(nItems, -2, 2)
-interceptsSml <- t(sapply(interceptsSml,
-                          function(x) x + seq(0.9, -0.9, length.out = 4)))
-colnames(interceptsSml) <- paste0("d", 1:ncol(interceptsSml))
-# uncorrelated traits
+slopesSml <- cbind(ci = rep(10, nItems),
+                   generate_slopes(nItems, sMSml[, 1, drop = FALSE],
+                                   FUN = rlnorm, meanlog = 0, sdlog = 0.3),
+                   generate_slopes(nItems, sMSml[, -1], 1))
+interceptsSml <- generate_intercepts(nItems, sMSml,
+                                     FUNd = runif, argsd = list(min = -2,
+                                                                max = 2),
+                                     FUNt = seq, argst = list(from = 0.9,
+                                                              to = -0.9,
+                                                              length.out = 4))
+# generating "subjects" - uncorrelated traits
 vcovTraitsSml <- matrix(0, nrow = ncol(sMSml) + 1, ncol = ncol(sMSml) + 1,
                         dimnames = list(c("ci", colnames(sMSml)),
                                         c("ci", colnames(sMSml))))
@@ -23,21 +24,16 @@ expectedTraitsSml <- setNames(c(2, rep(0, ncol(vcovTraitsSml) - 1)),
                               colnames(vcovTraitsSml))
 thetaSml = mnormt::rmnorm(50, mean = expectedTraitsSml, varcov = vcovTraitsSml)
 
+# generating items' parameters - "sequentaial" mode
 sMSqt <- make_scoring_matrix_aem(1:5, "mae")
-slopesSqt <- matrix(c(rep(10, nItems),
-                      rlnorm(nItems, 0, 0.2),
-                      rlnorm(nItems, 0, 0.2),
-                      rlnorm(nItems, 0, 0.2)),
-                    ncol = 4, dimnames = list(NULL, c("ci", colnames(sMSqt))))
-interceptsSqt <- matrix(c(rep(0, nItems),
-                          rnorm(nItems, 0, 1.5),
-                          rnorm(nItems, 0, 1.5),
-                          rnorm(nItems, 0, 1.5)),
-                        ncol = ncol(sMSqt) + 1,
-                        dimnames = list(NULL,
-                                        paste0(c("ci", colnames(sMSqt)), "1")))
-
-# uncorrelated traits
+slopesSqt <- cbind(ci = rep(10, nItems),
+                   generate_slopes(nItems, sMSqt,
+                                   FUN = rlnorm, meanlog = 0, sdlog = 0.2))
+interceptsSqt <- cbind(ci1 = rep(0, nItems),
+                       generate_intercepts(nItems, sMSqt,
+                                           FUNd = rnorm, argsd = list(mean = 0,
+                                                                      sd = 1.5)))
+# generating "subjects" - uncorrelated traits
 vcovTraitsSqt <- matrix(0, nrow = ncol(sMSqt) + 1, ncol = ncol(sMSqt) + 1,
                         dimnames = list(c("ci", colnames(sMSqt)),
                                         c("ci", colnames(sMSqt))))
@@ -48,13 +44,9 @@ expectedTraitsSqt <- setNames(c(2, rep(0, ncol(vcovTraitsSqt) - 1)),
 thetaSqt = mnormt::rmnorm(50, mean = expectedTraitsSqt, varcov = vcovTraitsSqt)
 ################################################################################
 # very intense straightlining with aem - simultaneous
-itemsSSml <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsSSml[[i]] <- make_item(sMSml, slopesSml[i, ], interceptsSml[i, ],
-                              "simultaneous",
-                              scoringOnPreviousResponses =
-                                score_on_last_answer_straight)
-}
+itemsSSml <- make_test(sMSml, slopesSml, interceptsSml, "simultaneous",
+                       scoringOnPreviousResponses =
+                         score_on_last_answer_straight)
 set.seed(26112020)
 respSSml <- generate_test_responses(thetaSml, itemsSSml)
 respSSml <- apply(respSSml, 1:2, as.numeric)
@@ -66,13 +58,9 @@ test_that("Straightlining can be generated with simultaneous A, M, E RS", {
 })
 ################################################################################
 # very intense straightlining with mae - sequential
-itemsSSqt <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsSSqt[[i]] <- make_item(sMSqt, slopesSqt[i, ], interceptsSqt[i, ],
-                              "sequential",
-                              scoringOnPreviousResponses =
-                                score_on_last_answer_straight)
-}
+itemsSSqt <- make_test(sMSqt, slopesSqt, interceptsSqt,  "sequential",
+                       scoringOnPreviousResponses =
+                         score_on_last_answer_straight)
 set.seed(26112020)
 respSSqt <- generate_test_responses(thetaSqt, itemsSSqt)
 respSSqt <- apply(respSSqt, 1:2, as.numeric)
@@ -84,13 +72,9 @@ test_that("Straightlining can be generated with sequential A, M, E RS", {
 })
 ################################################################################
 # very "bouncing" C/IRS with aem - simultaneous
-itemsBSml <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsBSml[[i]] <- make_item(sMSml, slopesSml[i, ], interceptsSml[i, ],
-                              "simultaneous",
-                              scoringOnPreviousResponses =
-                                score_on_previous_answers_bounce)
-}
+itemsBSml <- make_test(sMSml, slopesSml, interceptsSml, "simultaneous",
+                       scoringOnPreviousResponses =
+                         score_on_previous_answers_bounce)
 set.seed(26112020)
 respBSml <- generate_test_responses(thetaSml, itemsBSml)
 respBSml <- apply(respBSml, 1:2, as.numeric)
@@ -104,13 +88,9 @@ test_that("'Bouncing' pattern can be generated with simultaneous A, M, E RS", {
 })
 ################################################################################
 # very "bouncing" C/IRS with mae - sequential
-itemsBSqt <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsBSqt[[i]] <- make_item(sMSqt, slopesSqt[i, ], interceptsSqt[i, ],
-                              "sequential",
-                              scoringOnPreviousResponses =
-                                score_on_previous_answers_bounce)
-}
+itemsBSqt <- make_test(sMSqt, slopesSqt, interceptsSqt, "sequential",
+                       scoringOnPreviousResponses =
+                         score_on_previous_answers_bounce)
 set.seed(26112020)
 respBSqt <- generate_test_responses(thetaSqt, itemsBSqt)
 respBSqt <- apply(respBSqt, 1:2, as.numeric)
@@ -124,13 +104,9 @@ test_that("'Bouncing' pattern can be generated with sequential A, M, E RS", {
 })
 ################################################################################
 # very  'next answer' C/IRS with aem - simultaneous
-itemsNSml <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsNSml[[i]] <- make_item(sMSml, slopesSml[i, ], interceptsSml[i, ],
-                              "simultaneous",
-                              scoringOnPreviousResponses =
-                                score_on_last_answer_next)
-}
+itemsNSml <- make_test(sMSml, slopesSml, interceptsSml, "simultaneous",
+                       scoringOnPreviousResponses =
+                         score_on_last_answer_next)
 set.seed(26112020)
 respNSml <- generate_test_responses(thetaSml, itemsNSml)
 respNSml <- apply(respNSml, 1:2, as.numeric)
@@ -144,13 +120,9 @@ test_that("'Next answer' pattern can be generated with simultaneous A, M, E RS",
 })
 ################################################################################
 # very 'next answer' C/IRS with mae - sequential
-itemsNSqt <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsNSqt[[i]] <- make_item(sMSqt, slopesSqt[i, ], interceptsSqt[i, ],
-                              "sequential",
-                              scoringOnPreviousResponses =
-                                score_on_last_answer_next)
-}
+itemsNSqt <- make_test(sMSqt, slopesSqt, interceptsSqt, "sequential",
+                       scoringOnPreviousResponses =
+                         score_on_last_answer_next)
 set.seed(26112020)
 respNSqt <- generate_test_responses(thetaSqt, itemsNSqt)
 respNSqt <- apply(respNSqt, 1:2, as.numeric)
@@ -164,13 +136,9 @@ test_that("'Next answer' pattern can be generated with sequential A, M, E RS", {
 })
 ################################################################################
 # very 'previous answer' C/IRS with aem - simultaneous
-itemsPSml <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsPSml[[i]] <- make_item(sMSml, slopesSml[i, ], interceptsSml[i, ],
-                              "simultaneous",
-                              scoringOnPreviousResponses =
-                                score_on_last_answer_previous)
-}
+itemsPSml <- make_test(sMSml, slopesSml, interceptsSml, "simultaneous",
+                       scoringOnPreviousResponses =
+                         score_on_last_answer_previous)
 set.seed(26112020)
 respPSml <- generate_test_responses(thetaSml, itemsPSml)
 respPSml <- apply(respPSml, 1:2, as.numeric)
@@ -184,13 +152,9 @@ test_that("'Previous answer' pattern can be generated with simultaneous A, M, E 
 })
 ################################################################################
 # very "'previous answer' C/IRS with mae - sequential
-itemsPSqt <- vector(mode = "list", length = nItems)
-for (i in 1:nItems) {
-  itemsPSqt[[i]] <- make_item(sMSqt, slopesSqt[i, ], interceptsSqt[i, ],
-                              "sequential",
-                              scoringOnPreviousResponses =
-                                score_on_last_answer_previous)
-}
+itemsPSqt <- make_test(sMSqt, slopesSqt, interceptsSqt, "sequential",
+                       scoringOnPreviousResponses =
+                         score_on_last_answer_previous)
 set.seed(26112020)
 respPSqt <- generate_test_responses(thetaSqt, itemsPSqt)
 respPSqt <- apply(respPSqt, 1:2, as.numeric)