+++ NEWS.md
@@ -1,3 +1,15 @@
-# rstyles 0.2.0
+# rstyles 0.3.0 (5.05.2021)
+
+## New features
+
+- Functions that enables user to generate *items* and *tests*: `generate_slopes()`, `generate_intercepts()` and `make_test()`.
+- Functions to compute log-normal distribution parameters on the scale of the actual variable (i.e. the exponential scale): `lnorm_mean()`, `lnorm_sd()` and `find_pars_lnorm()`.
+- Function that enables user to generate a very simple *scoring matrix* that doesn't deal with any response styles: `make_scoring_matrix_trivial()`. It is useful if one wants to use *rstyles* for simulations that don't involve response styles.
+
+## Bug-fixes
+
+- Some improvements in documentation.
+
+# rstyles 0.2.0 (17.12.2020)
 
 - First ready to use version of the package.